clean up code into component files, add comments
clean up more code conflicts in App.js and CalendarScreen.js
