+++ tpp-app/src/home/pages/CalendarScreen.js
@@ -1,10 +1,10 @@
-import React, {useState} from 'react';
-import { View, StyleSheet, Text, Image, TouchableOpacity } from 'react-native';
+import React, { useState } from 'react';
+import { View, StyleSheet } from 'react-native';
 import { CalendarList } from 'react-native-calendars';
 import { BackButton } from '../components/BackButtonComponent';
 import Selector from '../components/Selector';
 import Icon from 'react-native-vector-icons/MaterialIcons';
-import {Button} from 'react-native-elements';
+import { Button } from 'react-native-elements';
 
 const VIEWS = {
     Flow: "Period Flow",
