+++ tpp-app/src/home/pages/SelectLogOptionOverlay.js
@@ -1,19 +1,8 @@
 import React from "react";
-import { Text, View, StyleSheet, TouchableWithoutFeedback, TouchableOpacity } from "react-native";
+import { View, StyleSheet, TouchableWithoutFeedback } from "react-native";
 import VectorImage from 'react-native-vector-image';
+import { OptionButton } from "../components/LoggingOptionButton";
 
-/**
- * Button component for different logging options
- */
-const OptionButton = ({ title, icon, onPress }) =>
-  (
-    <TouchableOpacity onPress={onPress} style={styles.button}>
-      <View style={styles.iconCircle}>
-        {icon}
-      </View>
-      <Text style={styles.optionText}>{title}</Text>
-    </TouchableOpacity>
-  )
 
 export default function SelectLogOptionOverlay({ navigation }) {
   return (
@@ -34,7 +23,7 @@ export default function SelectLogOptionOverlay({ navigation }) {
               title={"Log multiple period dates"}
               icon={<VectorImage source={require('../../../ios/tppapp/Images.xcassets/icons/red_calendar.svg')}/>}
               onPress={() => {
-                navigation.goBack(); // dismiss this overlay first
+                navigation.goBack();
                 navigation.navigate('LogMultipleDates');
               }}
             />
@@ -55,35 +44,6 @@ const styles = StyleSheet.create({
   buttonContainer: {
     position: 'absolute',
     bottom: 45,
-    justifyContent: 'center'
-  },
-  button: {
-    flexDirection: 'row',
-    justifyContent: 'center',
-    alignItems: 'center',
-    marginBottom: 21
-  },
-  optionText: {
-    textAlign: 'center',
-    maxWidth: 100,
-    marginLeft: 15,
-    fontSize: 14
-  },
-  iconCircle: {
-    backgroundColor: '#fff',
-    width: 54,
-    height: 54,
-    borderRadius: 54,
-    shadowColor: "#000",
-    shadowOffset: {
-      width: 0,
-      height: 4,
-    },
-    shadowOpacity: 0.35,
-    shadowRadius: 8,
-    elevation: 5,
-    display: 'flex',
-    alignItems: 'center',
     justifyContent: 'center'
   }
 });
